Refactored according to review comments
