+++ spring-cloud-aws-messaging/src/main/java/org/springframework/cloud/aws/messaging/listener/SimpleMessageListenerContainer.java
@@ -25,9 +25,7 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import com.amazonaws.handlers.AsyncHandler;
 import com.amazonaws.services.sqs.model.DeleteMessageRequest;
-import com.amazonaws.services.sqs.model.DeleteMessageResult;
 import com.amazonaws.services.sqs.model.Message;
 import com.amazonaws.services.sqs.model.ReceiveMessageResult;
 
@@ -42,6 +40,7 @@ import static org.springframework.cloud.aws.messaging.core.QueueMessageUtils.cre
 /**
  * @author Agim Emruli
  * @author Alain Sahli
+ * @author Mete Alpaslan Katircioglu
  * @since 1.0
  */
 public class SimpleMessageListenerContainer extends AbstractMessageListenerContainer {
@@ -443,22 +442,7 @@ public class SimpleMessageListenerContainer extends AbstractMessageListenerConta
 		private void deleteMessage(String receiptHandle) {
 			getAmazonSqs().deleteMessageAsync(
 					new DeleteMessageRequest(this.queueUrl, receiptHandle),
-					new AsyncHandler<DeleteMessageRequest, DeleteMessageResult>() {
-						@Override
-						public void onError(Exception exception) {
-							getLogger().warn(
-									"An exception occurred while deleting '{}' receiptHandle",
-									receiptHandle, exception);
-						}
-
-						@Override
-						public void onSuccess(DeleteMessageRequest request,
-								DeleteMessageResult deleteMessageResult) {
-							getLogger().trace(
-									"'{}' receiptHandle is deleted successfully",
-									request.getReceiptHandle());
-						}
-					});
+					new DeleteMessageHandler(receiptHandle));
 		}
 
 		private org.springframework.messaging.Message<String> getMessageForExecution() {
